discordListener/python/requirements.txt
python-dotenv
PyNaCl
pydub
google-cloud-storage
gspread
oauth2client
flask
flask_restful